pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mysql-connector-wrapper"
version = "0.1.0"
description = "A wrapper for mysql-connector-python"
requires-python = ">=3.6"
dependencies = ["mysql-connector-python>=8.0.0"]
